Hooks in CleanupService for all ResourceIterators
o Ensures that IAGD dependency resolver only performs correct casts
o Javadocs for ResourceIter*
o Makes Node.getLabels return a ResourceIterable
o Makes IndexHits ResourceIter*

+++ community/kernel/src/main/java/org/neo4j/kernel/impl/cleanup/CleanupReference.java
@@ -29,32 +29,44 @@ class CleanupReference extends PhantomReference<Object>
     private Closeable handler;
     CleanupReference prev, next;
 
-    CleanupReference(Object referent, ReferenceQueueBasedCleanupService cleanupService, Closeable handler)
+    CleanupReference( Object referent, ReferenceQueueBasedCleanupService cleanupService, Closeable handler )
     {
         super( referent, cleanupService.collectedReferences.queue );
         this.cleanupService = cleanupService;
         this.handler = handler;
     }
 
-    void cleanupNow( boolean explicit ) throws IOException {
-        synchronized ( this )
+    void cleanupNow( boolean explicit ) throws IOException
+    {
+        boolean shouldUnlink = false;
+        try
         {
-            if ( handler != null )
+            synchronized ( this )
             {
-                if (!explicit)
+                if ( handler != null )
                 {
-                    cleanupService.logLeakedReference(this);
-                }
-                try
-                {
-                    handler.close();
-                }
-                finally
-                {
-                    handler = null;
+                    shouldUnlink = true;
+                    if ( !explicit )
+                    {
+                        cleanupService.logLeakedReference( this );
+                    }
+                    try
+                    {
+                        handler.close();
+                    }
+                    finally
+                    {
+                        handler = null;
+                    }
                 }
             }
         }
-        cleanupService.unlink( this );
+        finally
+        {
+            if ( shouldUnlink )
+            {
+                cleanupService.unlink( this );
+            }
+        }
     }
 }
